+++ CHANGELOG.md
@@ -4,8 +4,9 @@ Version 1 is implemented without Docker. (Git Clone and systemctl service) which
 
 There is no versioning before 2.0.0 so the changelog is not here.
 
-## [Unreleased - patch]
+## [Unreleased - minor]
 
+- feat: live Discord status for compose restarts; continue unless two consecutive failures
 - chore: bump dependencies
 
 ## [3.1.1] - 2026-04-05